Copy files in stage 1 using specified credentials
Stage one of the publish/install deployment was not using the remote
credentials in order to connect to the target machine to copy the script
and package files.  This resulted in the caller needing to have access
(usually admin) on the target machine.  This change allows deployment
regardless of the one executing the cmdlets as long as the
-RemoteCredential has the required access to the remote share and
filesystem.