Enable running of post install script on target
Although one could remote to a target machine after running Powerdeploy,
including the ability within Powerdeploy (perhaps as a switch) allows an
easy way to connect to the machine and run the script in a user
context that is already configured and working (rather than needing to
work through the connection intricacies of PowerShell Remoting.  Use of
deployment context variables (extraction folders, etc.) which might be
needed (with a cleanup script, for example) are available.

Resolves issue #4