Remove 'v' from built zip and root at Powerdeploy
The Powerdeploy zip file built as part of the build process was adding 'v'
to the version number for the filename which doesn't match a common
convention for uploaded file names in Bitbucket (App-1.2.3.zip).  Also,
the zip file had the module files in the root of the zip which made it
hard to extract to module directory because you have to rename the folder
that gets extracted.  This change allows the user to drag and drop the
Powerdeploy folder to their modules directory.