Store package script registrations in context
Powerdeploy was storing various state in different top-level $global:pd...
variables.  These are being consolidated to live under the deployment
context variable to limit the surface area of variables and reduce the
possibility of conflicts.

In order to facilitate this, new functions were added that allow anything
running in the install side of the deployment to store state in the
deployment context using Set-DeploymentContextState and subsequently load
it with Get-DeploymentContextState.