Provide extended error info from remote execution
When a deployment fails on a remote target, the error information
previously returned was not very helpful.  The only information available
for troubleshooting was the error message itself with no context into the
running script or line on the remote.

This information is now provided as verbose output if the remote execution
fails.

The session and package cleanup will also run now even if a remote error
occurs, rather than just terminating the script immediately.