Start acceptance tests of fully loaded module
All of the tests for the project up to this point simulate a module
loading up to speed up tests and isolate components.  The acceptance tests
provide end-to-end tests of use cases to verify that the module works as
expected as a whole.  After a duplicate Get-DeploymentVariable messed up
custom installation scripts in a test (it's loaded in Powerdeploy and
Installer modules), it was apparent we needed tests of the loaded module.